// ExistingCode/src/data-access/domain/tables/assaybatch/schema.ts
export const AssayBatchTableInsertBaseSchema = z.object({
	/**
	 * dataType: uniqueidentifier
	 * defaultValue: (newid())
	 */
	assayBatchId: z.any().optional(),
	/**
	 * dataType: nvarchar
	 */
	labCode: z.string().max(20),
	/**
	 * dataType: nvarchar
	 */
	batchNo: z.string().max(50),
	/**
	 * dataType: nvarchar
	 */
	dispatchNo: z.string().max(50).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	dispatchedDt: z.any().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	labReceivedDt: z.any().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	labPrelimDt: z.any().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	labJobDt: z.any().nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	labFinalDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	labSamplePrep: z.string().max(255).nullable().optional(),
	/**
	 * dataType: bit
	 * defaultValue: ((0))
	 */
	validated: z.any().optional(),
	/**
	 * dataType: datetimeoffset
	 */
	validatedDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	sourceFile: z.string().max(255).nullable().optional(),
	/**
	 * dataType: int
	 */
	sampleCount: z.number().int().nullable().optional(),
	/**
	 * dataType: int
	 */
	samplesMerged: z.number().int().nullable().optional(),
	/**
	 * dataType: int
	 */
	resultsCount: z.number().int().nullable().optional(),
	/**
	 * dataType: int
	 */
	resultsMerged: z.number().int().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	contractId: z.string().max(20).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	mergeDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: ((-1))
	 */
	batchStatus: z.string().max(50).optional(),
	/**
	 * dataType: nvarchar
	 */
	comments: z.string().max(1000).nullable().optional(),
	/**
	 * dataType: datetimeoffset
	 * defaultValue: (sysdatetimeoffset())
	 */
	createdOnDt: z.any().optional(),
	/**
	 * dataType: nvarchar
	 * defaultValue: (suser_sname())
	 */
	createdBy: z.string().max(50).optional(),
	/**
	 * dataType: datetimeoffset
	 */
	modifiedOnDt: z.any().nullable().optional(),
	/**
	 * dataType: nvarchar
	 */
	modifiedBy: z.string().max(50).nullable().optional(),
	/**
	 * dataType: timestamp
	 */
	rv: z.date(),
});